requirements.txt
streamlit
requests
google-generativeai
streamlit_ace
uuid
datetime
